Refactor move i18n middleware in I18n module
