+++ src/modules/I18n/next-i18next.config.js
